+++ src/components/DashBoard/css/tabs.module.css
@@ -32,7 +32,7 @@
     border-style: none;
     border-color: none;
     border-image: initial;
-    margin-top: 4rem;
+    margin-top: 1rem;
 }
 
 .tabs::selection{
