+++ src/components/DashBoard/dashboard.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import NavBar from "../../Common/navigation";
+import NavBar from "../Common/navigation";
 import UserCard from "./userCard";
 // import DashboardTab from "./dashboardTab";
 import clsx from "clsx";
