+++ src/pages/halodev.js
@@ -2,9 +2,8 @@ import Home from '../components/Home/homepage';
 import Footer from '../components/Common/footer';
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Compete from "../components/compete/compete";
-import styles from './halodev.module.css';
 import CompeteDetail from '../components/compete/competeDetail/competeDetail';
-import DashBoard from '../components/compete/DashBoard/dashboard';
+import DashBoard from '../components/DashBoard/dashboard';
 import 'bootstrap/dist/css/bootstrap.min.css';
